Move FakeRunner into fake folder
Also:
- refator switch localtion
