+++ switch_operations/switch.go
@@ -1,4 +1,4 @@
-package switch_interface
+package switch_operations
 
 type Switch interface {
 	Update(string) error
